+++ src/rules/__tests__/no-deprecated-colors.test.js
@@ -28,14 +28,16 @@ const ruleTester = new RuleTester({
 ruleTester.run('no-deprecated-colors', rule, {
   valid: [
     `import {Box} from '@other/design-system'; <Box color="text.primary">Hello</Box>`,
-    `import {Box} from "@primer/components"; <Box color="fg.default">Hello</Box>`,
-    `import {hello} from "@primer/components"; hello("colors.text.primary")`,
-    `import {themeGet} from "@primer/components"; themeGet("space.text.primary")`,
+    `import {Box} from "@primer/react"; <Box color="fg.default">Hello</Box>`,
+    `import {hello} from "@primer/react"; hello("colors.text.primary")`,
+    `import {themeGet} from "@primer/react"; themeGet("space.text.primary")`,
+    `import {themeGet} from "@primer/react"; themeGet(props.backgroundColorThemeValue)`,
+    `import {themeGet} from "@primer/react"; themeGet(2)`,
     `import {themeGet} from "@other/design-system"; themeGet("colors.text.primary")`,
     `import {get} from "@other/constants"; get("space.text.primary")`,
-    `import {Box} from '@primer/components'; <Box sx={styles}>Hello</Box>`,
-    `import {Box} from '@primer/components'; <Box sx={{color: text.primary}}>Hello</Box>`,
-    `import {Box} from '@primer/components'; <Box sx={{color: "fg.default"}}>Hello</Box>`,
+    `import {Box} from '@primer/react'; <Box sx={styles}>Hello</Box>`,
+    `import {Box} from '@primer/react'; <Box sx={{color: text.primary}}>Hello</Box>`,
+    `import {Box} from '@primer/react'; <Box sx={{color: "fg.default"}}>Hello</Box>`,
     `{color: 'text.primary'}`
   ],
   invalid: [
@@ -50,8 +52,8 @@ ruleTester.run('no-deprecated-colors', rule, {
       ]
     },
     {
-      code: `import {Box} from "@primer/components"; function Example() { return <Box color="text.primary">Hello</Box> }`,
-      output: `import {Box} from "@primer/components"; function Example() { return <Box color="fg.default">Hello</Box> }`,
+      code: `import {Box} from "@primer/react"; function Example() { return <Box color="text.primary">Hello</Box> }`,
+      output: `import {Box} from "@primer/react"; function Example() { return <Box color="fg.default">Hello</Box> }`,
       errors: [
         {
           message: '"text.primary" is deprecated. Use "fg.default" instead.'
@@ -69,8 +71,8 @@ ruleTester.run('no-deprecated-colors', rule, {
       ]
     },
     {
-      code: `import Box from '@primer/components/lib-esm/Box'; function Example() { return <Box color="text.primary">Hello</Box> }`,
-      output: `import Box from '@primer/components/lib-esm/Box'; function Example() { return <Box color="fg.default">Hello</Box> }`,
+      code: `import Box from '@primer/react/lib-esm/Box'; function Example() { return <Box color="text.primary">Hello</Box> }`,
+      output: `import Box from '@primer/react/lib-esm/Box'; function Example() { return <Box color="fg.default">Hello</Box> }`,
       errors: [
         {
           message: '"text.primary" is deprecated. Use "fg.default" instead.'
@@ -78,8 +80,8 @@ ruleTester.run('no-deprecated-colors', rule, {
       ]
     },
     {
-      code: `import {Box} from "@primer/components"; const Example = () => <Box color="text.primary">Hello</Box>`,
-      output: `import {Box} from "@primer/components"; const Example = () => <Box color="fg.default">Hello</Box>`,
+      code: `import {Box} from "@primer/react"; const Example = () => <Box color="text.primary">Hello</Box>`,
+      output: `import {Box} from "@primer/react"; const Example = () => <Box color="fg.default">Hello</Box>`,
       errors: [
         {
           message: '"text.primary" is deprecated. Use "fg.default" instead.'
@@ -87,8 +89,8 @@ ruleTester.run('no-deprecated-colors', rule, {
       ]
     },
     {
-      code: `import {Box} from "@primer/components"; <Box bg="bg.primary" m={1} />`,
-      output: `import {Box} from "@primer/components"; <Box bg="canvas.default" m={1} />`,
+      code: `import {Box} from "@primer/react"; <Box bg="bg.primary" m={1} />`,
+      output: `import {Box} from "@primer/react"; <Box bg="canvas.default" m={1} />`,
       errors: [
         {
           message: '"bg.primary" is deprecated. Use "canvas.default" instead.'
@@ -96,8 +98,8 @@ ruleTester.run('no-deprecated-colors', rule, {
       ]
     },
     {
-      code: `import {Box} from "@primer/components"; <Box sx={{bg: "bg.primary", m: 1, ...rest}} />`,
-      output: `import {Box} from "@primer/components"; <Box sx={{bg: "canvas.default", m: 1, ...rest}} />`,
+      code: `import {Box} from "@primer/react"; <Box sx={{bg: "bg.primary", m: 1, ...rest}} />`,
+      output: `import {Box} from "@primer/react"; <Box sx={{bg: "canvas.default", m: 1, ...rest}} />`,
       errors: [
         {
           message: '"bg.primary" is deprecated. Use "canvas.default" instead.'
@@ -105,8 +107,8 @@ ruleTester.run('no-deprecated-colors', rule, {
       ]
     },
     {
-      code: `import {Box} from "@primer/components"; <Box sx={{boxShadow: theme => theme.shadows.autocomplete.shadow}} />`,
-      output: `import {Box} from "@primer/components"; <Box sx={{boxShadow: theme => theme.shadows.shadow.medium}} />`,
+      code: `import {Box} from "@primer/react"; <Box sx={{boxShadow: theme => theme.shadows.autocomplete.shadow}} />`,
+      output: `import {Box} from "@primer/react"; <Box sx={{boxShadow: theme => theme.shadows.shadow.medium}} />`,
       errors: [
         {
           message: '"theme.shadows.autocomplete.shadow" is deprecated. Use "theme.shadows.shadow.medium" instead.'
@@ -114,8 +116,8 @@ ruleTester.run('no-deprecated-colors', rule, {
       ]
     },
     {
-      code: `import {Box} from "@primer/components"; <Box sx={{boxShadow: theme => \`0 1px 2px \${theme.colors.text.primary}\`}} />`,
-      output: `import {Box} from "@primer/components"; <Box sx={{boxShadow: theme => \`0 1px 2px \${theme.colors.fg.default}\`}} />`,
+      code: `import {Box} from "@primer/react"; <Box sx={{boxShadow: theme => \`0 1px 2px \${theme.colors.text.primary}\`}} />`,
+      output: `import {Box} from "@primer/react"; <Box sx={{boxShadow: theme => \`0 1px 2px \${theme.colors.fg.default}\`}} />`,
       errors: [
         {
           message: '"theme.colors.text.primary" is deprecated. Use "theme.colors.fg.default" instead.'
@@ -123,8 +125,8 @@ ruleTester.run('no-deprecated-colors', rule, {
       ]
     },
     {
-      code: `import {Box} from "@primer/components"; <Box sx={{boxShadow: t => \`0 1px 2px \${t.colors.text.primary}\`}} />`,
-      output: `import {Box} from "@primer/components"; <Box sx={{boxShadow: t => \`0 1px 2px \${t.colors.fg.default}\`}} />`,
+      code: `import {Box} from "@primer/react"; <Box sx={{boxShadow: t => \`0 1px 2px \${t.colors.text.primary}\`}} />`,
+      output: `import {Box} from "@primer/react"; <Box sx={{boxShadow: t => \`0 1px 2px \${t.colors.fg.default}\`}} />`,
       errors: [
         {
           message: '"t.colors.text.primary" is deprecated. Use "t.colors.fg.default" instead.'
@@ -132,8 +134,8 @@ ruleTester.run('no-deprecated-colors', rule, {
       ]
     },
     {
-      code: `import {Box} from "@primer/components"; <Box sx={{"&:hover": {bg: "bg.primary"}}} />`,
-      output: `import {Box} from "@primer/components"; <Box sx={{"&:hover": {bg: "canvas.default"}}} />`,
+      code: `import {Box} from "@primer/react"; <Box sx={{"&:hover": {bg: "bg.primary"}}} />`,
+      output: `import {Box} from "@primer/react"; <Box sx={{"&:hover": {bg: "canvas.default"}}} />`,
       errors: [
         {
           message: '"bg.primary" is deprecated. Use "canvas.default" instead.'
@@ -141,25 +143,25 @@ ruleTester.run('no-deprecated-colors', rule, {
       ]
     },
     {
-      code: `import {Box} from "@primer/components"; <Box color="auto.green.5" />`,
+      code: `import {Box} from "@primer/react"; <Box color="auto.green.5" />`,
       errors: [
         {
           message: '"auto.green.5" is deprecated.',
           suggestions: [
             {
               desc: 'Use "success.fg" instead.',
-              output: `import {Box} from "@primer/components"; <Box color="success.fg" />`
+              output: `import {Box} from "@primer/react"; <Box color="success.fg" />`
             },
             {
               desc: 'Use "success.emphasis" instead.',
-              output: `import {Box} from "@primer/components"; <Box color="success.emphasis" />`
+              output: `import {Box} from "@primer/react"; <Box color="success.emphasis" />`
             }
           ]
         }
       ]
     },
     {
-      code: `import {Box} from "@primer/components"; <Box color="fade.fg10" />`,
+      code: `import {Box} from "@primer/react"; <Box color="fade.fg10" />`,
       errors: [
         {
           message:
@@ -168,8 +170,8 @@ ruleTester.run('no-deprecated-colors', rule, {
       ]
     },
     {
-      code: `import {Box, Text} from "@primer/components"; <Box bg="bg.primary"><Text color="text.primary">Hello</Text></Box>`,
-      output: `import {Box, Text} from "@primer/components"; <Box bg="canvas.default"><Text color="fg.default">Hello</Text></Box>`,
+      code: `import {Box, Text} from "@primer/react"; <Box bg="bg.primary"><Text color="text.primary">Hello</Text></Box>`,
+      output: `import {Box, Text} from "@primer/react"; <Box bg="canvas.default"><Text color="fg.default">Hello</Text></Box>`,
       errors: [
         {
           message: '"bg.primary" is deprecated. Use "canvas.default" instead.'
@@ -180,8 +182,8 @@ ruleTester.run('no-deprecated-colors', rule, {
       ]
     },
     {
-      code: `import {themeGet} from "@primer/components"; themeGet("colors.text.primary")`,
-      output: `import {themeGet} from "@primer/components"; themeGet("colors.fg.default")`,
+      code: `import {themeGet} from "@primer/react"; themeGet("colors.text.primary")`,
+      output: `import {themeGet} from "@primer/react"; themeGet("colors.fg.default")`,
       errors: [
         {
           message: '"colors.text.primary" is deprecated. Use "colors.fg.default" instead.'
@@ -189,8 +191,8 @@ ruleTester.run('no-deprecated-colors', rule, {
       ]
     },
     {
-      code: `import {themeGet} from "@primer/components"; themeGet("shadows.autocomplete.shadow")`,
-      output: `import {themeGet} from "@primer/components"; themeGet("shadows.shadow.medium")`,
+      code: `import {themeGet} from "@primer/react"; themeGet("shadows.autocomplete.shadow")`,
+      output: `import {themeGet} from "@primer/react"; themeGet("shadows.shadow.medium")`,
       errors: [
         {
           message: '"shadows.autocomplete.shadow" is deprecated. Use "shadows.shadow.medium" instead.'